+++ CHANGELOG.md
@@ -12,6 +12,7 @@ app.
   ([#978](https://github.com/birchill/10ten-ja-reader/issues/978)).
 - Made non-matching kanji headwords be hidden when matching on kana
   ([#979](https://github.com/birchill/10ten-ja-reader/issues/979)).
+- Handle search-only dictionary headwords.
 - Fixed display of currency conversion etc. when there are only name results
   ([#970](https://github.com/birchill/10ten-ja-reader/issues/970)).